ktlo(common): inline alloy-compat impls into types
