+++ crates/common/consensus/src/transaction/deposit.rs
@@ -13,6 +13,11 @@ use alloy_rlp::{BufMut, Decodable, Encodable, Header};
 
 use super::OpTxType;
 
+#[cfg(feature = "alloy-compat")]
+use alloy_network::{UnknownTxEnvelope, UnknownTypedTransaction};
+#[cfg(feature = "alloy-compat")]
+use alloy_rpc_types_eth::ConversionError;
+
 /// Deposit transactions, also known as deposits are initiated on L1, and executed on L2.
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Default)]
 #[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
@@ -329,6 +334,30 @@ impl Decodable for TxDeposit {
 impl Sealable for TxDeposit {
     fn hash_slow(&self) -> B256 {
         self.tx_hash()
+    }
+}
+
+#[cfg(feature = "alloy-compat")]
+impl TryFrom<UnknownTxEnvelope> for TxDeposit {
+    type Error = ConversionError;
+
+    fn try_from(value: UnknownTxEnvelope) -> Result<Self, Self::Error> {
+        value.inner.try_into()
+    }
+}
+
+#[cfg(feature = "alloy-compat")]
+impl TryFrom<UnknownTypedTransaction> for TxDeposit {
+    type Error = ConversionError;
+
+    fn try_from(value: UnknownTypedTransaction) -> Result<Self, Self::Error> {
+        if !value.is_type(OpTxType::Deposit as u8) {
+            return Err(ConversionError::Custom("invalid transaction type".to_string()));
+        }
+        value
+            .fields
+            .deserialize_into()
+            .map_err(|_| ConversionError::Custom("invalid transaction data".to_string()))
     }
 }
